+++ app/containers/sidebar.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import prompt from 'electron-prompt'
 
-import { viewChannel, joinChannel, changeUsername, changeScreen, showCabalSettings } from '../actions'
+import { viewChannel, joinChannel, changeUsername, changeScreen, hideCabalSettings, showCabalSettings } from '../actions'
 import Avatar from './avatar'
 
 const mapStateToProps = state => {
@@ -11,6 +11,7 @@ const mapStateToProps = state => {
     addr: state.currentCabal,
     cabals: state.cabals,
     cabal,
+    cabalSettingsVisible: state.cabalSettingsVisible,
     channelMessagesUnread: cabal.channelMessagesUnread,
     username: cabal.username
   }
@@ -21,7 +22,8 @@ const mapDispatchToProps = dispatch => ({
   joinChannel: ({ addr, channel }) => dispatch(joinChannel({ addr, channel })),
   viewChannel: ({ addr, channel }) => dispatch(viewChannel({ addr, channel })),
   changeUsername: ({ addr, username }) => dispatch(changeUsername({ addr, username })),
-  showCabalSettings: ({ key }) => dispatch(showCabalSettings({ key }))
+  showCabalSettings: ({ addr }) => dispatch(showCabalSettings({ addr })),
+  hideCabalSettings: () => dispatch(hideCabalSettings())
 })
 
 class SidebarScreen extends React.Component {
@@ -55,8 +57,12 @@ class SidebarScreen extends React.Component {
     })
   }
 
-  onClickCabalSettings (key) {
-    this.props.showCabalSettings({ key })
+  onClickCabalSettings (addr) {
+    if (this.props.cabalSettingsVisible) {
+      this.props.hideCabalSettings()
+    } else {
+      this.props.showCabalSettings({ addr })
+    }
   }
 
   joinChannel (channel) {
@@ -90,7 +96,7 @@ class SidebarScreen extends React.Component {
     return (
       <div className='client__sidebar' onClick={(e) => this.props.toggleEmojis(false)}>
         <div className='sidebar'>
-          <div className='session' onClick={self.onClickCabalSettings.bind(self, cabal.key)}>
+          <div className='session' onClick={self.onClickCabalSettings.bind(self, cabal.addr)}>
             <div className='session__avatar'>
               <div className='session__avatar__img'>
                 <Avatar name={username} />