+++ CHANGELOG.md
@@ -3,5 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [2.0.3] - 2019-06-18
 ### Added
+- UI to show and set channel topics.
+- Display currently connected peer count.
 - Settings screen for future features and a button to remove the cabal from the client.
-- Slash commands for /nick, /join (for switching channels), and /remove (for removing a cabal from the client).
+- Slash commands for /nick, /topic, /join (for switching channels), and /remove (for removing a cabal from the client).